Move await consensus commit functions into channel service
