+++ lib/collection/src/shards/channel_service.rs
@@ -1,8 +1,14 @@
 use std::collections::HashMap;
 use std::sync::Arc;
+use std::time::Duration;
 
+use api::grpc::qdrant::qdrant_internal_client::QdrantInternalClient;
+use api::grpc::qdrant::WaitOnConsensusCommitRequest;
 use api::grpc::transport_channel_pool::TransportChannelPool;
-use tonic::transport::Uri;
+use futures::future::try_join_all;
+use futures::Future;
+use tonic::transport::{Channel, Uri};
+use tonic::Status;
 use url::Url;
 
 use crate::operations::types::{CollectionError, CollectionResult};
@@ -32,6 +38,103 @@ impl ChannelService {
         if let Some(uri) = removed {
             self.channel_pool.drop_pool(&uri).await;
         }
+    }
+
+    /// Wait until all other known peers reach the given commit
+    ///
+    /// # Errors
+    ///
+    /// This errors if:
+    /// - any of the peers is not on the same term
+    /// - waiting takes longer than the specified timeout
+    /// - any of the peers cannot be reached
+    pub async fn await_commit_on_all_peers(
+        &self,
+        this_peer_id: PeerId,
+        commit: u64,
+        term: u64,
+        timeout: Duration,
+    ) -> Result<(), CollectionError> {
+        let requests = self
+            .id_to_address
+            .read()
+            .keys()
+            .filter(|id| **id != this_peer_id)
+            // The collective timeout at the bottom of this function handles actually timing out.
+            // Since an explicit timeout must be given here as well, it is multiplied by two to
+            // give the collective timeout some space.
+            .map(|peer_id| self.await_commit_on_peer(*peer_id, commit, term, timeout * 2))
+            .collect::<Vec<_>>();
+        let responses = try_join_all(requests);
+
+        // Handle requests with timeout
+        tokio::time::timeout(timeout, responses)
+            .await
+            .map(|_| ())
+            .map_err(|_elapsed| CollectionError::Timeout {
+                description: "Failed to wait for consensus commit on all peers, timed out.".into(),
+            })
+    }
+
+    /// Wait until the given peer reaches the given commit
+    ///
+    /// # Errors
+    ///
+    /// This errors if the given peer is on a different term. Also errors if the peer cannot be reached.
+    async fn await_commit_on_peer(
+        &self,
+        peer_id: PeerId,
+        commit: u64,
+        term: u64,
+        timeout: Duration,
+    ) -> Result<(), CollectionError> {
+        let response = self
+            .with_qdrant_client(peer_id, |mut client| async move {
+                let request = WaitOnConsensusCommitRequest {
+                    commit: commit as i64,
+                    term: term as i64,
+                    timeout: timeout.as_secs() as i64,
+                };
+                client
+                    .wait_on_consensus_commit(tonic::Request::new(request))
+                    .await
+            })
+            .await
+            .map_err(|err| {
+                CollectionError::service_error(format!(
+                    "Failed to wait for consensus commit on peer {peer_id}: {err}"
+                ))
+            })?
+            .into_inner();
+
+        // Create error if wait request failed
+        if !response.ok {
+            return Err(CollectionError::service_error(format!(
+                "Failed to wait for consensus commit on peer {peer_id}, has diverged commit/term or timed out."
+            )));
+        }
+        Ok(())
+    }
+
+    async fn with_qdrant_client<T, O: Future<Output = Result<T, Status>>>(
+        &self,
+        peer_id: PeerId,
+        f: impl Fn(QdrantInternalClient<Channel>) -> O,
+    ) -> Result<T, CollectionError> {
+        let address = self
+            .id_to_address
+            .read()
+            .get(&peer_id)
+            .ok_or_else(|| CollectionError::service_error("Address for peer ID is not found."))?
+            .clone();
+        self.channel_pool
+            .with_channel(&address, |channel| {
+                let client = QdrantInternalClient::new(channel);
+                let client = client.max_decoding_message_size(usize::MAX);
+                f(client)
+            })
+            .await
+            .map_err(Into::into)
     }
 
     /// Get the REST address for the current peer.
